+++ submitr/scripts/submit_metadata_bundle.py
@@ -111,6 +111,8 @@ ADVANCED OPTIONS:
   Displays ONLY all known consortia; nothing else.
 --submission-centers
   Displays ONLY all known submission centers; nothing else.
+--nouploads
+  Do not attempt to upload any files; use resume-uploads later.
 --json
   Displays the submitted metadata as formatted JSON.
 --json-only
@@ -164,8 +166,10 @@ def main(simulated_args_for_testing=None):
     parser.add_argument('--patch-only', action="store_true",
                         help="Only perform updates (PATCH) for submitted data.", default=False)
     parser.add_argument('--keys', help="Path to keys file (rather than default ~/.smaht-keys.json).", default=None)
-    parser.add_argument('--submit', action="store_true",
-                        help="Actually submit the metadata for ingestion..", default=False)
+    parser.add_argument('--submit-new', '--submit', action="store_true",
+                        help="Actually submit new metadata for ingestion; may NOT contain updtes.", default=False)
+    parser.add_argument('--submit-update', '--update', action="store_true",
+                        help="Actually submit the metadata for ingestion; may contain updates.", default=False)
     parser.add_argument('--validate', action="store_true",
                         help="Perform both client-side and server-side validation first.", default=False)
     parser.add_argument('--validate-local-only', action="store_true",
@@ -192,6 +196,8 @@ def main(simulated_args_for_testing=None):
                         help="Suppress (yes/no) requests for user input.", default=False)
     parser.add_argument('--ref-nocache', action="store_true",
                         help="Do not cache reference (linkTo) lookups.", default=False)
+    parser.add_argument('--nouploads', action="store_true",
+                        help="Do not attempt to upload any files; use resume-uploads later.", default=False)
     parser.add_argument('--noprogress', action="store_true",
                         help="Do not track progress of client-side parsing/validation.", default=False)
     parser.add_argument('--app',
@@ -356,13 +362,18 @@ def main(simulated_args_for_testing=None):
                              show_details=args.details,
                              post_only=args.post_only,
                              patch_only=args.patch_only,
-                             submit=args.submit,
+                             # The --submit option (aka --submit-new) allows items to
+                             # be created but not updated (at least not updates with diffs)
+                             submit=args.submit_new,
+                             # The --update option allow items to be created or updated.
+                             submit_update=args.submit_update,
                              rclone_google=cloud_store,
                              validate_local_only=args.validate_local_only,
                              validate_remote_only=args.validate_remote_only,
                              validate_local_skip=args.validate_local_skip,
                              validate_remote_skip=args.validate_remote_skip,
                              noanalyze=args.noanalyze,
+                             nouploads=args.nouploads,
                              json_only=args.json_only,
                              ref_nocache=args.ref_nocache,
                              merge=args.merge,
@@ -404,18 +415,22 @@ def _setup_validate_related_options(args: argparse.Namespace):
     if args.info:
         return
 
-    if args.submit:
+    if args.submit_new and args.submit_update:
+        PRINT(f"May not specify both --submit-new AND --update options.")
+        sys.exit(1)
+
+    if args.submit_new or args.submit_update:
         if args.validate or args.validate_local_only or args.validate_remote_only:
-            PRINT(f"May not specify both --submit AND validate options.")
+            PRINT(f"May not specify both --submit-new AND validate options.")
             sys.exit(1)
         if args.json_only:
-            PRINT(f"May not specify both --submit AND --json-only options.")
+            PRINT(f"May not specify both --submit-new AND --json-only options.")
             sys.exit(1)
 
-    if not args.submit:
+    if not (args.submit_new or args.submit_update):
         if not (args.validate or args.validate_local_only or args.validate_remote_only):
             if not args.json_only and not _pytesting():
-                PRINT(f"Must specify either --validate or --submit options. Use --help for all options.")
+                PRINT(f"Must specify either --validate or --submit-new options. Use --help for all options.")
                 sys.exit(1)
         elif args.json_only:
             PRINT(f"May not specify both --json-only and validate options.")
@@ -438,8 +453,8 @@ def _setup_validate_related_options(args: argparse.Namespace):
         sys.exit(1)
 
     if args.json_only:
-        if args.submit:
-            PRINT("The --json-only option is not allowed with --submit.")
+        if args.submit_new or args.submit_update:
+            PRINT("The --json-only option is not allowed with --submit-new or --update.")
             sys.exit(1)
         if args.validate_remote_only or args.validate_local_skip:
             PRINT("The --json-only option is not allowed with these validate options.")